test(omo-senpi): classify memorian upstream failures from the settled child turn
